requirements.txt
fastapi==0.104.1
openai==0.28.1
pymongo==3.11.0
python-dotenv==1.0.0
Requests==2.28.1
tiktoken==0.5.1
uvicorn==0.24.0.post1
